telegram.bot library is updated to v17

+++ BotLib/Engine/AdminTasks/ChatmemberInfoAdminTask.cs
@@ -5,12 +5,12 @@ namespace BotLib.Engine.AdminTasks
 {
     public class ChatmemberInfoAdminTask : AdminTask
     {
-        public ChatmemberInfoAdminTask(long ChatId, int UserId) : base(ChatId, AdminTaskType.ChatmemberInfo)
+        public ChatmemberInfoAdminTask(long ChatId, long UserId) : base(ChatId, AdminTaskType.ChatmemberInfo)
         {
             this.UserId = UserId;
         }
 
-        public ChatmemberInfoAdminTask(long ChatId, int UserId, DateTime DelayUntil) : base(ChatId, AdminTaskType.ChatmemberInfo, DelayUntil)
+        public ChatmemberInfoAdminTask(long ChatId, long UserId, DateTime DelayUntil) : base(ChatId, AdminTaskType.ChatmemberInfo, DelayUntil)
         {
             this.UserId = UserId;
         }
@@ -18,7 +18,7 @@ namespace BotLib.Engine.AdminTasks
         public string FirstName { get; private set; }
         public string LastName { get; private set; }
         public ChatMemberStatus Status { get; private set; }
-        public int UserId { get; }
+        public long UserId { get; }
         public string Username { get; private set; }
 
         internal void FillUserInfo(string FirstName, string LastName, string Username, ChatMemberStatus Status)
