+++ BotLib/Engine/AdminTasks/UnrestrictChatmemberAdminTask.cs
@@ -4,16 +4,16 @@ namespace BotLib.Engine.AdminTasks
 {
     public class UnrestrictChatmemberAdminTask : AdminTask
     {
-        public UnrestrictChatmemberAdminTask(long ChatId, int UserId) : base(ChatId, AdminTaskType.UnrestrictChatmember)
+        public UnrestrictChatmemberAdminTask(long ChatId, long UserId) : base(ChatId, AdminTaskType.UnrestrictChatmember)
         {
             this.UserId = UserId;
         }
 
-        public UnrestrictChatmemberAdminTask(long ChatId, int UserId, DateTime DelayUntil) : base(ChatId, AdminTaskType.UnrestrictChatmember, DelayUntil)
+        public UnrestrictChatmemberAdminTask(long ChatId, long UserId, DateTime DelayUntil) : base(ChatId, AdminTaskType.UnrestrictChatmember, DelayUntil)
         {
             this.UserId = UserId;
         }
 
-        public int UserId { get; }
+        public long UserId { get; }
     }
 }
