+++ BotLib/Engine/Commands/TelegramBlobCommand.cs
@@ -5,7 +5,7 @@ namespace BotLib.Engine.Commands
 {
     public abstract class TelegramBlobCommand : TelegramCommand
     {
-        public TelegramBlobCommand(TelegramCommandType CommandType, int UserId, object Argument) : base(CommandType, UserId, Argument)
+        public TelegramBlobCommand(TelegramCommandType CommandType, long UserId, object Argument) : base(CommandType, UserId, Argument)
         {
             if (Argument.GetType() != typeof(byte[]))
                 throw new InvalidCommandArgumentException("Should be byte[]", Argument.GetType());
