+++ BotLib/Engine/Commands/TelegramButtonPressedCommand.cs
@@ -4,7 +4,7 @@ namespace BotLib.Engine.Commands
 {
     public class TelegramButtonPressedCommand : TelegramCommand
     {
-        public TelegramButtonPressedCommand(int UserId, string CallbackData) : base(TelegramCommandType.ButtonPressed, UserId, CallbackData)
+        public TelegramButtonPressedCommand(long UserId, string CallbackData) : base(TelegramCommandType.ButtonPressed, UserId, CallbackData)
         {
         }
