+++ BotLib/Engine/Commands/TelegramGeoCommand.cs
@@ -4,13 +4,13 @@ namespace BotLib.Engine.Commands
 {
     public class TelegramGeoCommand : TelegramCommand
     {
-        public TelegramGeoCommand(int UserId, object Argument) : base(TelegramCommandType.Geo, UserId, Argument)
+        public TelegramGeoCommand(long UserId, object Argument) : base(TelegramCommandType.Geo, UserId, Argument)
         {
-            if (Argument.GetType() != typeof(float[])) throw new InvalidCommandArgumentException("Invalid type of command's argument", Argument.GetType());
-            if (((float[])Argument).Length != 2) throw new InvalidCommandArgumentException("The length of array is not 2", typeof(float[]));
+            if (Argument.GetType() != typeof(double[])) throw new InvalidCommandArgumentException("Invalid type of command's argument", Argument.GetType());
+            if (((double[])Argument).Length != 2) throw new InvalidCommandArgumentException("The length of array is not 2", typeof(double[]));
         }
 
-        public float Latitude => ((float[])Argument)[0];
-        public float Longitude => ((float[])Argument)[1];
+        public double Latitude => ((double[])Argument)[0];
+        public double Longitude => ((double[])Argument)[1];
     }
 }
