+++ BotLib/Engine/Commands/TelegramTextCommand.cs
@@ -5,7 +5,7 @@ namespace BotLib.Engine.Commands
 {
     public class TelegramTextCommand : TelegramCommand
     {
-        public TelegramTextCommand(int UserId, object Argument) : base(TelegramCommandType.TextEntered, UserId, Argument)
+        public TelegramTextCommand(long UserId, object Argument) : base(TelegramCommandType.TextEntered, UserId, Argument)
         {
             if (Argument.GetType() != typeof(string)) throw new InvalidCommandArgumentException("Invalid type of command's argument", Argument.GetType());
         }
